move old key filtering tests into plan tests
